pull file and tag out of BasicResourceItem
